Refactoring, added database code, more ui for boards, global errors
Login routing is now within the `routes/authentication.js` file. Board
routing is new and unfinished, but it is in `routes/boards.js`.

Most of the database is now complete (I think...). Boards, Stories, and
Tasks all exist now.

Boards now have a create UI. It doesn't actually create them yet, but it
will eventually.

Global error handling can now be done on the front end. So far, 404
errors will take you to a new 404 page.

That's about it...