Did some more stuff with boards.
Users only see boards they are allowed to see. If they are not a member
of the board, it will not show up in their dashboard.

Viewing a board is now a thing. You can click the link from the
dashboard and go into a board view. You can't see or do much of
anything, but it is a start.