Finished the stories gui, started on the board gui.
The stories gui is done for now. There will probably be changes later,
but for now it's good. Descriptions, avatars and an active checkbox are
all now visiible.

Tasks can be made active or inactive, which will mark if they are shown
on the board or not. By default, they are not.

The board gui is now able to show all active tasks. It just shows the
descriptions, so nothing fancy yet.