Redid how stories and tasks are.
Made them far more coupled together. Instead of stories being in their
own tab, kind of unrelated to the board, they are now the cards. Tasks
are shown when the stories are clicked on and have a checkbox to say if
they are done or not.

The done/todo count is shown on the cards and when everything is
finished, it is given a green background to show it is ready to move on.

The boards also have hard coded in lists. Todo, Develop, Test and Done.
Cards should ideally start at Todo and make their way towards Done.

I think this design will be way simpler than what I had originally
thought of...