You can now add and change sprints.
You start off with one blank sprint with the start day of the day you
created the board. Then, you can create sprints with the option to
transfer stories that have not been completed yet.

The board will always load with the first sprint that the current date
falls in between. Afterwards, you can increment or decrement the sprint
with two buttons.