Did some refactoring and made it a bit more fluent to use.
Sprints are no longer sent along with the board. This caused all the
sprints to be sent along, and that could get network intensive at some
point.

Sprints can now be accessed in the url. You can type in
/#/boards/id/indexOfSprint, and that will take you to a sprint (assuming
it exists). The buttons that navagate through sprints use this method
now instead of using JavaScript to change it.