You can now remove/delete stories.
If they are removed from a story, a prompt will appear asking the user
if they are sure they want to delete it. When they click yes, it will
remove the story from the sprint. If the story doesn't exist in any
other sprints, it will also delete it from the database entirely.
Otherwise, it will keep it in the database, so it will stay on
previous/later sprints.